requirements.txt
django
python-dotenv
python-telegram-bot
apscheduler
gunicorn
whitenoise
